requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.1.14
django-3-jet==1.0.8
django-crispy-forms==1.14.0
django-import-export==2.7.1
django-unused-media==0.2.2
et-xmlfile==1.1.0
future==0.18.2
gunicorn==20.1.0
heroku==0.1.4
html5lib==1.1
idna==3.3
MarkupPy==1.14
npm==0.1.1
odfpy==1.4.1
openpyxl==3.0.9
optional-django==0.1.0
Pillow==8.4.0
psycopg2==2.9.3
pycparser==2.20
PyMySQL==1.0.2
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==1.5
pytz==2021.3
PyYAML==6.0
reportlab==3.6.6
requests==2.26.0
six==1.16.0
source==1.2.0
sqlparse==0.4.2
tablib==3.2.0
tkintertable==1.3.3
touch==2020.12.3
typing_extensions==4.0.1
tzdata==2021.5
urllib3==1.26.7
values==2020.12.3
webencodings==0.5.1
whitenoise==5.3.0
xhtml2pdf==0.2.5
xlrd==2.0.1
xlwt==1.3.0
